renderIndex moved to main.go
